requirements.txt
keyboard==0.13.5
pywinauto==0.6.8
pywin32>=307
comtypes>=1.2.0
pystray>=0.19.0
Pillow>=10.0.0
